+++ auto_route/example/lib/mobile/screens/books/book_list_page.dart
@@ -9,10 +9,7 @@ class BookListScreen extends StatefulWidget {
   State<BookListScreen> createState() => _BookListScreenState();
 }
 
-class _BookListScreenState extends State<BookListScreen>
-    with AutoRouteAwareStateMixin<BookListScreen> {
-
-
+class _BookListScreenState extends State<BookListScreen> with AutoRouteAwareStateMixin<BookListScreen> {
   @override
   void didPushNext() {
     print('didPushNext');
@@ -22,26 +19,19 @@ class _BookListScreenState extends State<BookListScreen>
   Widget build(BuildContext context) {
     var booksDb = BooksDBProvider.of(context);
     return Scaffold(
-      body: ListView(
-        children: booksDb?.books
-                .map((book) => Column(
-                      children: [
-                        Card(
-                          margin: const EdgeInsets.symmetric(
-                              horizontal: 16, vertical: 8),
-                          child: ListTile(
-                            title: Text(book.name),
-                            subtitle: Text(book.genre),
-                            onTap: () {
-                              context.pushRoute(BookDetailsRoute(id: book.id));
-                            },
-                          ),
-                        ),
-                      ],
-                    ))
-                .toList() ??
-            const [],
-      ),
+      body: ListView(children: [
+        for (final book in [...?booksDb?.books])
+          Card(
+            margin: const EdgeInsets.symmetric(horizontal: 16, vertical: 8),
+            child: ListTile(
+              title: Text(book.name),
+              subtitle: Text(book.genre),
+              onTap: () {
+                context.pushRoute(BookDetailsRoute(id: book.id));
+              },
+            ),
+          ),
+      ]),
     );
   }
 }