+++ auto_route/lib/src/route/route_data.dart
@@ -6,7 +6,6 @@ class RouteData {
   final RoutingController router;
   final RouteType type;
   final TitleBuilder? _titleBuilder;
-
   LocalKey get key => _match.key;
 
   RouteData({