+++ auto_route/lib/src/router/controller/auto_router_delegate.dart
@@ -103,12 +103,13 @@ class AutoRouterDelegate extends RouterDelegate<UrlState> with ChangeNotifier {
   void _onNewUrlState(UrlState state) {
     final pathInBrowser = state.uri.path;
     var matchedUrlState = state.flatten;
-
     if (pathInBrowser != matchedUrlState.path) {
       matchedUrlState = matchedUrlState.copyWith(shouldReplace: true);
     }
     controller.navigationHistory.onNewUrlState(matchedUrlState);
   }
+
+
 
   @override
   Widget build(BuildContext context) => _AutoRootRouter(