+++ auto_route/lib/src/router/controller/routing_controller.dart
@@ -22,14 +22,21 @@ part 'root_stack_router.dart';
 typedef RouteDataPredicate = bool Function(RouteData route);
 typedef OnNestedNavigateCallBack = void Function(List<RouteMatch> routes);
 typedef OnTabNavigateCallBack = void Function(RouteMatch route);
-typedef RoutesBuilder = List<PageRouteInfo> Function(
-    PendingRoutesHandler handler);
+typedef RoutesBuilder = List<PageRouteInfo> Function(PendingRoutesHandler handler);
 typedef RoutePopCallBack = void Function(RouteMatch route, dynamic results);
 typedef OnNavigateCallBack = void Function(UrlState tree);
 typedef NavigatorObserversBuilder = List<NavigatorObserver> Function();
 
 abstract class RoutingController with ChangeNotifier {
   final _childControllers = <RoutingController>[];
+  bool _ignorePopCompleter = false;
+
+  bool get ignorePopCompleters => root._ignorePopCompleter;
+
+  @visibleForTesting
+  set ignorePopCompleters(bool value) {
+    root._ignorePopCompleter = value;
+  }
 
   List<RoutingController> get childControllers => _childControllers;
   final List<AutoRoutePage> _pages = [];
@@ -49,8 +56,7 @@ abstract class RoutingController with ChangeNotifier {
     _childControllers.remove(childController);
   }
 
-  RoutingController? _innerControllerOf(Key? key) =>
-      _childControllers.lastWhereOrNull(
+  RoutingController? _innerControllerOf(Key? key) => _childControllers.lastWhereOrNull(
         (c) => c.key == key,
       );
 
@@ -79,8 +85,67 @@ abstract class RoutingController with ChangeNotifier {
     }
   }
 
-  List<RouteData> get stackData =>
-      List.unmodifiable(_pages.map((e) => e.routeData));
+  List<HierarchySegment> currentHierarchy({
+    bool asPath = false,
+    bool ignorePending = false,
+    bool ignoreParams = false,
+  }) =>
+      _getCurrentHierarchy(
+        stackData,
+        asPath: asPath,
+        ignoreParams: ignoreParams,
+        ignorePending: ignorePending,
+      );
+
+  List<HierarchySegment> _getCurrentHierarchy(
+    List<RouteData> stack, {
+    required bool asPath,
+    required bool ignorePending,
+    required bool ignoreParams,
+  }) {
+    final hierarchy = <HierarchySegment>[];
+    for (final data in stack) {
+      final segmentName = asPath ? data.match : data.name;
+      final childCtrl = _innerControllerOf(data.key);
+      final childSegments = <HierarchySegment>[];
+      if (childCtrl?.hasEntries == true) {
+        childSegments.addAll(_getCurrentHierarchy(
+          childCtrl!.stackData,
+          asPath: asPath,
+          ignorePending: ignorePending,
+          ignoreParams: ignoreParams,
+        ));
+      } else if (!ignorePending && data.hasPendingChildren) {
+        childSegments.addAll(
+          _getCurrentHierarchy(
+            [
+              for (final pendingRoute in data.pendingChildren)
+                RouteData(
+                  route: pendingRoute,
+                  router: data.router,
+                  pendingChildren: [...?pendingRoute.children],
+                  type: data.type,
+                ),
+            ],
+            asPath: asPath,
+            ignorePending: ignorePending,
+            ignoreParams: ignoreParams,
+          ),
+        );
+      }
+      hierarchy.add(
+        HierarchySegment(
+          segmentName,
+          children: childSegments,
+          pathParams: ignoreParams || data.pathParams.isEmpty ? null : data.pathParams,
+          queryParams: ignoreParams || data.queryParams.isEmpty ? null : data.queryParams,
+        ),
+      );
+    }
+    return hierarchy;
+  }
+
+  List<RouteData> get stackData => List.unmodifiable(_pages.map((e) => e.routeData));
 
   bool isRouteActive(String routeName) {
     return navigationHistory.isRouteActive(routeName);
@@ -122,8 +187,7 @@ abstract class RoutingController with ChangeNotifier {
     if (match != null) {
       return match;
     } else {
-      final paths =
-          routeCollection.findPathTo(route.routeName).map((e) => e.name);
+      final paths = routeCollection.findPathTo(route.routeName).map((e) => e.name);
       final path = paths.isEmpty ? '' : paths.reduce((a, b) => a += ' -> $b');
       if (onFailure != null) {
         onFailure(RouteNotFoundFailure(path));
@@ -158,12 +222,9 @@ abstract class RoutingController with ChangeNotifier {
     return routeCollection.containsKey(route.routeName);
   }
 
-  _RouterScopeResult<T>?
-      _findPathScopeOrReportFailure<T extends RoutingController>(String path,
-          {bool includePrefixMatches = false, OnNavigationFailure? onFailure}) {
-    final routers = topMostRouter(ignorePagelessRoutes: true)
-        ._buildRoutersHierarchy()
-        .whereType<T>();
+  _RouterScopeResult<T>? _findPathScopeOrReportFailure<T extends RoutingController>(String path,
+      {bool includePrefixMatches = false, OnNavigationFailure? onFailure}) {
+    final routers = topMostRouter(ignorePagelessRoutes: true)._buildRoutersHierarchy().whereType<T>();
 
     for (var router in routers) {
       final matches = router.matcher.match(
@@ -184,25 +245,20 @@ abstract class RoutingController with ChangeNotifier {
     return null;
   }
 
-  RoutingController _findScope<T extends RoutingController>(
-      PageRouteInfo route) {
-    return topMostRouter(ignorePagelessRoutes: true)
-        ._buildRoutersHierarchy()
-        .firstWhere(
+  RoutingController _findScope<T extends RoutingController>(PageRouteInfo route) {
+    return topMostRouter(ignorePagelessRoutes: true)._buildRoutersHierarchy().firstWhere(
           (r) => r._canHandleNavigation(route),
           orElse: () => this,
         );
   }
 
-  Future<dynamic> navigate(PageRouteInfo route,
-      {OnNavigationFailure? onFailure}) async {
+  Future<dynamic> navigate(PageRouteInfo route, {OnNavigationFailure? onFailure}) async {
     return _findScope(route)._navigate(route, onFailure: onFailure);
   }
 
   void _onNavigate(List<RouteMatch> routes);
 
-  Future<dynamic> _navigate(PageRouteInfo route,
-      {OnNavigationFailure? onFailure}) async {
+  Future<dynamic> _navigate(PageRouteInfo route, {OnNavigationFailure? onFailure}) async {
     final match = _matchOrReportFailure(route, onFailure);
     if (match != null) {
       return _navigateAll([match], onFailure: onFailure);
@@ -230,8 +286,7 @@ abstract class RoutingController with ChangeNotifier {
   }
 
   List<RoutingController> _buildRoutersHierarchy() {
-    void collectRouters(
-        RoutingController currentParent, List<RoutingController> all) {
+    void collectRouters(RoutingController currentParent, List<RoutingController> all) {
       all.add(currentParent);
       if (currentParent._parent != null) {
         collectRouters(currentParent._parent!, all);
@@ -270,13 +325,14 @@ abstract class RoutingController with ChangeNotifier {
 
   bool get canNavigateBack => navigationHistory.canNavigateBack;
 
- @Deprecated('use back() instead')
+  @Deprecated('use back() instead')
   void navigateBack() => navigationHistory.back();
+
   void back() => navigationHistory.back();
 
-  void pushPathState(Object? state)=> navigationHistory.pushPathState(state);
-  Object? get pathState => navigationHistory.pathState;
+  void pushPathState(Object? state) => navigationHistory.pushPathState(state);
 
+  Object? get pathState => navigationHistory.pathState;
 
   RootStackRouter get root => (_parent?.root ?? this) as RootStackRouter;
 
@@ -321,8 +377,7 @@ abstract class RoutingController with ChangeNotifier {
   Future<bool> pop<T extends Object?>([T? result]);
 
   @optionalTypeArgs
-  Future<bool> popTop<T extends Object?>([T? result]) =>
-      topMostRouter().pop<T>(result);
+  Future<bool> popTop<T extends Object?>([T? result]) => topMostRouter().pop<T>(result);
 
   @Deprecated('Use canPop instead')
   bool get canPopSelfOrChildren;
@@ -350,29 +405,20 @@ abstract class RoutingController with ChangeNotifier {
     return segments;
   }
 
-  PageRouteInfo? buildPageRoute(String? path,
-      {bool includePrefixMatches = true}) {
+  PageRouteInfo? buildPageRoute(String? path, {bool includePrefixMatches = true}) {
     if (path == null) return null;
-    return matcher
-        .match(path, includePrefixMatches: includePrefixMatches)
-        ?.firstOrNull
-        ?.toPageRouteInfo();
+    return matcher.match(path, includePrefixMatches: includePrefixMatches)?.firstOrNull?.toPageRouteInfo();
   }
 
-  List<PageRouteInfo>? buildPageRoutesStack(String? path,
-      {bool includePrefixMatches = true}) {
+  List<PageRouteInfo>? buildPageRoutesStack(String? path, {bool includePrefixMatches = true}) {
     if (path == null) return null;
-    return matcher
-        .match(path, includePrefixMatches: includePrefixMatches)
-        ?.map((m) => m.toPageRouteInfo())
-        .toList();
+    return matcher.match(path, includePrefixMatches: includePrefixMatches)?.map((m) => m.toPageRouteInfo()).toList();
   }
 
   @override
   String toString() => '${routeData.name} Router';
 
-  Future<void> _navigateAll(List<RouteMatch> routes,
-      {OnNavigationFailure? onFailure});
+  Future<void> _navigateAll(List<RouteMatch> routes, {OnNavigationFailure? onFailure});
 }
 
 class TabsRouter extends RoutingController {
@@ -501,15 +547,13 @@ class TabsRouter extends RoutingController {
     }
   }
 
-  void replaceAll(
-      List<PageRouteInfo> routes, PageRouteInfo<dynamic> previousActiveRoute) {
+  void replaceAll(List<PageRouteInfo> routes, PageRouteInfo<dynamic> previousActiveRoute) {
     final routesToPush = _matchAllOrReportFailure(routes)!;
 
     _pages.clear();
     _childControllers.clear();
     _pushAll(routesToPush);
-    var targetIndex =
-        routesToPush.indexWhere((r) => r.name == previousActiveRoute.routeName);
+    var targetIndex = routesToPush.indexWhere((r) => r.name == previousActiveRoute.routeName);
     if (targetIndex == -1) {
       targetIndex = homeIndex == -1 ? 0 : homeIndex;
     }
@@ -517,8 +561,7 @@ class TabsRouter extends RoutingController {
   }
 
   @override
-  Future<void> _navigateAll(List<RouteMatch> routes,
-      {OnNavigationFailure? onFailure}) async {
+  Future<void> _navigateAll(List<RouteMatch> routes, {OnNavigationFailure? onFailure}) async {
     if (routes.isNotEmpty) {
       final mayUpdateRoute = routes.last;
 
@@ -546,8 +589,7 @@ class TabsRouter extends RoutingController {
         var mayUpdateController = _innerControllerOf(mayUpdateRoute.key);
         if (mayUpdateController != null) {
           final newRoutes = mayUpdateRoute.children ?? const [];
-          return mayUpdateController._navigateAll(newRoutes,
-              onFailure: onFailure);
+          return mayUpdateController._navigateAll(newRoutes, onFailure: onFailure);
         }
       }
       _updateSharedPathData(
@@ -621,8 +663,7 @@ class TabsRouter extends RoutingController {
       fragment: fragment,
     ));
     if (includeAncestors && _parent != null) {
-      _parent!
-          ._updateSharedPathData(queryParams: queryParams, fragment: fragment);
+      _parent!._updateSharedPathData(queryParams: queryParams, fragment: fragment);
     }
   }
 
@@ -657,8 +698,7 @@ abstract class StackRouter extends RoutingController {
   void _attachRedirectGuard(AutoRedirectGuardBase guard) {
     final stackRouters = _buildRoutersHierarchy().whereType<StackRouter>();
 
-    if (stackRouters
-        .any((r) => r._redirectGuardsListeners.containsKey(guard))) {
+    if (stackRouters.any((r) => r._redirectGuardsListeners.containsKey(guard))) {
       return;
     }
 
@@ -709,9 +749,7 @@ abstract class StackRouter extends RoutingController {
     if (_pages.length > 1 || hasPagelessTopRoute) {
       return true;
     } else if (_pages.isNotEmpty) {
-      return _innerControllerOf(_pages.last.routeData.key)
-              ?.canPopSelfOrChildren ??
-          false;
+      return _innerControllerOf(_pages.last.routeData.key)?.canPopSelfOrChildren ?? false;
     }
     return false;
   }
@@ -787,8 +825,7 @@ abstract class StackRouter extends RoutingController {
 
   @override
   RoutingController topMostRouter({bool ignorePagelessRoutes = false}) {
-    if (_childControllers.isNotEmpty &&
-        (ignorePagelessRoutes || !hasPagelessTopRoute)) {
+    if (_childControllers.isNotEmpty && (ignorePagelessRoutes || !hasPagelessTopRoute)) {
       var topRouteKey = currentChild?.key;
       final innerRouter = _innerControllerOf(topRouteKey);
       if (innerRouter != null) {
@@ -893,23 +930,19 @@ abstract class StackRouter extends RoutingController {
   List<AutoRoutePage> get stack => List.unmodifiable(_pages);
 
   @optionalTypeArgs
-  Future<T?> push<T extends Object?>(PageRouteInfo route,
-      {OnNavigationFailure? onFailure}) async {
+  Future<T?> push<T extends Object?>(PageRouteInfo route, {OnNavigationFailure? onFailure}) async {
     return _findStackScope(route)._push<T>(route, onFailure: onFailure);
   }
 
   StackRouter _findStackScope(PageRouteInfo route) {
-    final stackRouters = topMostRouter(ignorePagelessRoutes: true)
-        ._buildRoutersHierarchy()
-        .whereType<StackRouter>();
+    final stackRouters = topMostRouter(ignorePagelessRoutes: true)._buildRoutersHierarchy().whereType<StackRouter>();
     return stackRouters.firstWhere(
       (c) => c._canHandleNavigation(route),
       orElse: () => this,
     );
   }
 
-  Future<dynamic> _popUntilOrPushAll(List<RouteMatch> routes,
-      {OnNavigationFailure? onFailure}) async {
+  Future<dynamic> _popUntilOrPushAll(List<RouteMatch> routes, {OnNavigationFailure? onFailure}) async {
     final anchor = routes.first;
     final anchorPage = _pages.lastWhereOrNull(
       (p) => p.routeKey == anchor.key,
@@ -1110,8 +1143,7 @@ abstract class StackRouter extends RoutingController {
       if (await _canNavigate(
         route,
         onFailure,
-        pendingRoutes:
-            routes.whereIndexed((index, element) => index > i).toList(),
+        pendingRoutes: routes.whereIndexed((index, element) => index > i).toList(),
       )) {
         if (i != (routes.length - 1)) {
           _addEntry(route, notify: false);
@@ -1137,6 +1169,11 @@ abstract class StackRouter extends RoutingController {
     RouteMatch route, {
     bool notify = true,
   }) {
+    final topRoute = _pages.lastOrNull?.routeData;
+    if (topRoute != null && topRoute._match.keepHistory == false) {
+      markUrlStateForReplace();
+      _removeRoute(topRoute._match, notify: false);
+    }
     final data = _createRouteData(route, routeData);
     final page = pageBuilder(data);
     _pages.add(page);