+++ auto_route/lib/src/router/parser/route_information_parser.dart
@@ -1,7 +1,8 @@
 import 'package:auto_route/auto_route.dart';
 import 'package:collection/collection.dart';
 import 'package:flutter/foundation.dart';
-import 'package:flutter/widgets.dart' show RouteInformation, RouteInformationParser;
+import 'package:flutter/widgets.dart'
+    show RouteInformation, RouteInformationParser;
 import 'package:path/path.dart' as p;
 
 import '../../matcher/route_matcher.dart';
@@ -15,8 +16,10 @@ class DefaultRouteParser extends RouteInformationParser<UrlState> {
   @override
   Future<UrlState> parseRouteInformation(RouteInformation routeInformation) {
     final uri = Uri.parse(routeInformation.location ?? '');
-    var matches = _matcher.matchUri(uri, includePrefixMatches: includePrefixMatches);
-    return SynchronousFuture<UrlState>(UrlState(uri, matches ?? const [],pathState: routeInformation.state));
+    var matches =
+        _matcher.matchUri(uri, includePrefixMatches: includePrefixMatches);
+    return SynchronousFuture<UrlState>(
+        UrlState(uri, matches ?? const [], pathState: routeInformation.state));
   }
 
   @override
@@ -95,11 +98,18 @@ class UrlState {
     }
   }
 
+
+  @override
+  String toString() {
+    return 'UrlState{uri: $uri, shouldReplace: $shouldReplace, pathState: $pathState}';
+  }
+
   bool get hasSegments => segments.isNotEmpty;
 
   RouteMatch? get topMatch => hasSegments ? segments.last : null;
 
-  UrlState get flatten => UrlState.fromSegments(segments.last.flattened,state: pathState);
+  UrlState get flatten =>
+      UrlState.fromSegments(segments.last.flattened, state: pathState);
 
   RouteMatch? _findSegment(
     List<RouteMatch> segments,
@@ -118,8 +128,11 @@ class UrlState {
     return null;
   }
 
+
   List<RouteMatch> childrenOfSegmentNamed(String routeName) {
-    return _findSegment(segments, (match) => match.name == routeName)?.children ?? const [];
+    return _findSegment(segments, (match) => match.name == routeName)
+            ?.children ??
+        const [];
   }
 
   static Uri _buildUri(List<RouteMatch> routes) {