+++ auto_route/lib/src/router/widgets/auto_tabs_router.dart
@@ -5,14 +5,12 @@ import 'package:flutter/material.dart';
 
 import '../../../auto_route.dart';
 
-typedef AnimatedIndexedStackBuilder = Widget Function(
+typedef AnimatedIndexedStackBuilder = Widget Function(BuildContext context, Widget child);
+typedef AnimatedIndexedStackTransitionBuilder = Widget Function(
     BuildContext context, Widget child, Animation<double> animation);
-typedef AutoTabsBuilder = Widget Function(
-    BuildContext context, List<Widget> children, TabsRouter tabsRouter);
-typedef AutoTabsPageViewBuilder = Widget Function(
-    BuildContext context, Widget child, PageController pageController);
-typedef AutoTabsTabBarBuilder = Widget Function(
-    BuildContext context, Widget child, TabController tabController);
+typedef AutoTabsBuilder = Widget Function(BuildContext context, List<Widget> children, TabsRouter tabsRouter);
+typedef AutoTabsPageViewBuilder = Widget Function(BuildContext context, Widget child, PageController pageController);
+typedef AutoTabsTabBarBuilder = Widget Function(BuildContext context, Widget child, TabController tabController);
 typedef OnNavigationChanged = Function(TabsRouter tabsRouter);
 
 abstract class AutoTabsRouter extends StatefulWidget {
@@ -30,8 +28,7 @@ abstract class AutoTabsRouter extends StatefulWidget {
     required this.routes,
     this.homeIndex = -1,
     this.inheritNavigatorObservers = true,
-    this.navigatorObservers =
-        AutoRouterDelegate.defaultNavigatorObserversBuilder,
+    this.navigatorObservers = AutoRouterDelegate.defaultNavigatorObserversBuilder,
   }) : super(key: key);
 
   const factory AutoTabsRouter({
@@ -41,6 +38,7 @@ abstract class AutoTabsRouter extends StatefulWidget {
     Duration duration,
     Curve curve,
     AnimatedIndexedStackBuilder? builder,
+    AnimatedIndexedStackTransitionBuilder? transitionBuilder,
     int homeIndex,
     bool inheritNavigatorObservers,
     NavigatorObserversBuilder navigatorObservers,
@@ -156,6 +154,7 @@ abstract class _AutoTabsRouterState extends State<AutoTabsRouter> {
 // -----------------------------------------------------------
 class _AutoTabsRouterIndexedStack extends AutoTabsRouter {
   final AnimatedIndexedStackBuilder? builder;
+  final AnimatedIndexedStackTransitionBuilder? transitionBuilder;
   final Duration duration;
   final Curve curve;
   final bool lazyLoad;
@@ -167,10 +166,10 @@ class _AutoTabsRouterIndexedStack extends AutoTabsRouter {
     this.duration = const Duration(milliseconds: 300),
     this.curve = Curves.ease,
     this.builder,
+    this.transitionBuilder,
     int homeIndex = -1,
     bool inheritNavigatorObservers = true,
-    NavigatorObserversBuilder navigatorObservers =
-        AutoRouterDelegate.defaultNavigatorObserversBuilder,
+    NavigatorObserversBuilder navigatorObservers = AutoRouterDelegate.defaultNavigatorObserversBuilder,
   }) : super._(
           key: key,
           routes: routes,
@@ -180,19 +179,16 @@ class _AutoTabsRouterIndexedStack extends AutoTabsRouter {
         );
 
   @override
-  _AutoTabsRouterIndexedStackState createState() =>
-      _AutoTabsRouterIndexedStackState();
+  _AutoTabsRouterIndexedStackState createState() => _AutoTabsRouterIndexedStackState();
 }
 
-class _AutoTabsRouterIndexedStackState extends _AutoTabsRouterState
-    with SingleTickerProviderStateMixin {
+class _AutoTabsRouterIndexedStackState extends _AutoTabsRouterState with SingleTickerProviderStateMixin {
   late AnimationController _animationController;
   late Animation<double> _animation;
   int _index = 0;
   late int _tabsHash;
 
-  _AutoTabsRouterIndexedStack get typedWidget =>
-      widget as _AutoTabsRouterIndexedStack;
+  _AutoTabsRouterIndexedStack get typedWidget => widget as _AutoTabsRouterIndexedStack;
 
   @override
   void initState() {
@@ -275,15 +271,23 @@ class _AutoTabsRouterIndexedStackState extends _AutoTabsRouterState
       child: TabsRouterScope(
         controller: _controller!,
         stateHash: stateHash,
-        child: AnimatedBuilder(
-          animation: _animation,
-          builder: (context, child) => builder(
+        child: Builder(builder: (context) {
+          if (typedWidget.transitionBuilder == null) {
+            return builderChild;
+          }
+          return builder(
             context,
-            child ?? builderChild,
-            _animation,
-          ),
-          child: builderChild,
-        ),
+            AnimatedBuilder(
+              animation: _animation,
+              child: builderChild,
+              builder: (context, child) => typedWidget.transitionBuilder!(
+                context,
+                child!,
+                _animation,
+              ),
+            ),
+          );
+        }),
       ),
     );
   }
@@ -317,14 +321,12 @@ class _IndexedStackBuilder extends StatefulWidget {
   _IndexedStackBuilderState createState() => _IndexedStackBuilderState();
 }
 
-class _IndexedStackBuilderState extends State<_IndexedStackBuilder>
-    with _RouteAwareTabsMixin<_IndexedStackBuilder> {
+class _IndexedStackBuilderState extends State<_IndexedStackBuilder> with _RouteAwareTabsMixin<_IndexedStackBuilder> {
   final _dummyWidget = const SizedBox.shrink();
   final _initializedPagesTracker = <int, bool>{};
 
   @override
-  List<RouteMatch> get routes =>
-      widget.stack.map((e) => e.routeData.route).toList();
+  List<RouteMatch> get routes => widget.stack.map((e) => e.routeData.route).toList();
 
   @override
   List<NavigatorObserver> get observers => widget.navigatorObservers;
@@ -354,8 +356,7 @@ class _IndexedStackBuilderState extends State<_IndexedStackBuilder>
       _setup();
       return;
     }
-    if (widget.lazyLoad &&
-        _initializedPagesTracker[widget.activeIndex] != true) {
+    if (widget.lazyLoad && _initializedPagesTracker[widget.activeIndex] != true) {
       _initializedPagesTracker[widget.activeIndex] = true;
       _didInitTabRoute(widget.activeIndex, oldWidget.activeIndex);
     } else if (widget.activeIndex != oldWidget.activeIndex) {
@@ -372,14 +373,11 @@ class _IndexedStackBuilderState extends State<_IndexedStackBuilder>
       children: List.generate(
         widget.stack.length,
         (index) {
-          if (!widget.stack[index].maintainState &&
-              index != widget.activeIndex) {
+          if (!widget.stack[index].maintainState && index != widget.activeIndex) {
             _initializedPagesTracker[index] = false;
           }
           final isInitialized = _initializedPagesTracker[index] == true;
-          return isInitialized
-              ? widget.itemBuilder(context, index)
-              : _dummyWidget;
+          return isInitialized ? widget.itemBuilder(context, index) : _dummyWidget;
         },
       ),
     );
@@ -407,8 +405,7 @@ class AutoTabsRouterPageView extends AutoTabsRouter {
     this.physics,
     this.dragStartBehavior = DragStartBehavior.start,
     bool inheritNavigatorObservers = true,
-    NavigatorObserversBuilder navigatorObservers =
-        AutoRouterDelegate.defaultNavigatorObserversBuilder,
+    NavigatorObserversBuilder navigatorObservers = AutoRouterDelegate.defaultNavigatorObserversBuilder,
   })  : _pageViewModeBuilder = builder,
         super._(
           key: key,
@@ -422,8 +419,7 @@ class AutoTabsRouterPageView extends AutoTabsRouter {
   AutoTabsRouterPageViewState createState() => AutoTabsRouterPageViewState();
 }
 
-class AutoTabsRouterPageViewState extends _AutoTabsRouterState
-    with _RouteAwareTabsMixin<AutoTabsRouter> {
+class AutoTabsRouterPageViewState extends _AutoTabsRouterState with _RouteAwareTabsMixin<AutoTabsRouter> {
   late PageController _pageController;
 
   @override
@@ -440,8 +436,7 @@ class AutoTabsRouterPageViewState extends _AutoTabsRouterState
         controllerPage = 0;
       }
       if (_controller!.activeIndex != controllerPage) {
-        _didChangeTabRoute(
-            _controller!.activeIndex, _controller!.previousIndex!);
+        _didChangeTabRoute(_controller!.activeIndex, _controller!.previousIndex!);
       }
     });
   }
@@ -456,8 +451,7 @@ class AutoTabsRouterPageViewState extends _AutoTabsRouterState
   void didUpdateWidget(covariant AutoTabsRouterPageView oldWidget) {
     super.didUpdateWidget(oldWidget);
     if (!const ListEquality().equals(widget.routes, oldWidget.routes)) {
-      _controller!.replaceAll(
-          widget.routes, oldWidget.routes[_controller!.activeIndex]);
+      _controller!.replaceAll(widget.routes, oldWidget.routes[_controller!.activeIndex]);
       _updatePageController();
     }
   }
@@ -502,8 +496,7 @@ class AutoTabsRouterPageViewState extends _AutoTabsRouterState
   List<NavigatorObserver> get observers => _navigatorObservers;
 
   @override
-  List<RouteMatch> get routes =>
-      _controller!.stackData.map((e) => e.route).toList();
+  List<RouteMatch> get routes => _controller!.stackData.map((e) => e.route).toList();
 }
 
 class _AutoTabsRouterTabBar extends AutoTabsRouter {
@@ -513,6 +506,7 @@ class _AutoTabsRouterTabBar extends AutoTabsRouter {
   final ScrollPhysics? physics;
   final DragStartBehavior dragStartBehavior;
   final Axis scrollDirection;
+
   const _AutoTabsRouterTabBar({
     Key? key,
     required List<PageRouteInfo> routes,
@@ -522,8 +516,7 @@ class _AutoTabsRouterTabBar extends AutoTabsRouter {
     this.duration,
     this.curve = Curves.ease,
     bool inheritNavigatorObservers = true,
-    NavigatorObserversBuilder navigatorObservers =
-        AutoRouterDelegate.defaultNavigatorObserversBuilder,
+    NavigatorObserversBuilder navigatorObservers = AutoRouterDelegate.defaultNavigatorObserversBuilder,
     this.physics,
     this.dragStartBehavior = DragStartBehavior.start,
   }) : super._(
@@ -581,8 +574,7 @@ class _AutoTabsRouterTabBarState extends _AutoTabsRouterState
   void didUpdateWidget(covariant _AutoTabsRouterTabBar oldWidget) {
     super.didUpdateWidget(oldWidget);
     if (!const ListEquality().equals(widget.routes, oldWidget.routes)) {
-      _controller!
-          .replaceAll(widget.routes, oldWidget.routes[_tabController.index]);
+      _controller!.replaceAll(widget.routes, oldWidget.routes[_tabController.index]);
       _updateTabController();
     }
   }
@@ -627,8 +619,7 @@ class _AutoTabsRouterTabBarState extends _AutoTabsRouterState
   List<NavigatorObserver> get observers => _navigatorObservers;
 
   @override
-  List<RouteMatch> get routes =>
-      _controller!.stackData.map((e) => e.route).toList();
+  List<RouteMatch> get routes => _controller!.stackData.map((e) => e.route).toList();
 }
 
 class _AutoTabsRouterBuilder extends AutoTabsRouter {
@@ -644,8 +635,7 @@ class _AutoTabsRouterBuilder extends AutoTabsRouter {
     required this.builder,
     int homeIndex = -1,
     bool inheritNavigatorObservers = true,
-    NavigatorObserversBuilder navigatorObservers =
-        AutoRouterDelegate.defaultNavigatorObserversBuilder,
+    NavigatorObserversBuilder navigatorObservers = AutoRouterDelegate.defaultNavigatorObserversBuilder,
   }) : super._(
           key: key,
           routes: routes,
@@ -658,8 +648,7 @@ class _AutoTabsRouterBuilder extends AutoTabsRouter {
   _AutoTabsRouterBuilderState createState() => _AutoTabsRouterBuilderState();
 }
 
-class _AutoTabsRouterBuilderState extends _AutoTabsRouterState
-    with _RouteAwareTabsMixin<AutoTabsRouter> {
+class _AutoTabsRouterBuilderState extends _AutoTabsRouterState with _RouteAwareTabsMixin<AutoTabsRouter> {
   @override
   void _setupController() {
     assert(_controller != null);
@@ -668,8 +657,7 @@ class _AutoTabsRouterBuilderState extends _AutoTabsRouterState
     _didInitTabRoute(_controller!.activeIndex);
     _controller!.addListener(() {
       if (_controller!.activeIndex != _controller!.previousIndex) {
-        _didChangeTabRoute(
-            _controller!.activeIndex, _controller!.previousIndex ?? 0);
+        _didChangeTabRoute(_controller!.activeIndex, _controller!.previousIndex ?? 0);
         typedWidget.onNavigate?.call(_controller!);
       }
       if (mounted) {
@@ -682,8 +670,7 @@ class _AutoTabsRouterBuilderState extends _AutoTabsRouterState
   void didUpdateWidget(covariant _AutoTabsRouterBuilder oldWidget) {
     super.didUpdateWidget(oldWidget);
     if (!const ListEquality().equals(widget.routes, oldWidget.routes)) {
-      _controller!.replaceAll(
-          widget.routes, oldWidget.routes[_controller!.activeIndex]);
+      _controller!.replaceAll(widget.routes, oldWidget.routes[_controller!.activeIndex]);
       typedWidget.onNavigate?.call(_controller!);
     }
   }
@@ -725,8 +712,7 @@ class _AutoTabsRouterBuilderState extends _AutoTabsRouterState
   List<NavigatorObserver> get observers => _navigatorObservers;
 
   @override
-  List<RouteMatch> get routes =>
-      _controller!.stackData.map((e) => e.route).toList();
+  List<RouteMatch> get routes => _controller!.stackData.map((e) => e.route).toList();
 }
 
 mixin _RouteAwareTabsMixin<T extends StatefulWidget> on State<T> {
@@ -738,8 +724,7 @@ mixin _RouteAwareTabsMixin<T extends StatefulWidget> on State<T> {
     observers.whereType<AutoRouterObserver>().forEach((observer) {
       TabPageRoute? previousRoute;
       if (previous != -1) {
-        previousRoute =
-            TabPageRoute(routeInfo: routes[previous], index: previous);
+        previousRoute = TabPageRoute(routeInfo: routes[previous], index: previous);
       }
       observer.didInitTabRoute(
         TabPageRoute(routeInfo: routes[index], index: index),
@@ -769,8 +754,7 @@ class KeepAliveTab extends StatefulWidget {
   State<KeepAliveTab> createState() => _KeepAliveTabState();
 }
 
-class _KeepAliveTabState extends State<KeepAliveTab>
-    with AutomaticKeepAliveClientMixin {
+class _KeepAliveTabState extends State<KeepAliveTab> with AutomaticKeepAliveClientMixin {
   @override
   Widget build(BuildContext context) {
     super.build(context);