+++ auto_route/test/nested_router/router_test.dart
@@ -7,13 +7,11 @@ import 'router.dart';
 
 void main() {
   late NestedRouter router;
-
   setUp(() {
     router = NestedRouter();
   });
 
-  testWidgets(
-      'Pushing ${SecondHostRoute.name} with no children should present [$SecondNested1Route]',
+  testWidgets('Pushing ${SecondHostRoute.name} with no children should present [$SecondNested1Route]',
       (WidgetTester tester) async {
     await pumpRouterApp(tester, router);
     router.push(const SecondHostRoute());
@@ -32,12 +30,12 @@ void main() {
     expect(router.urlState.url, '/second');
   });
 
-  testWidgets(
-      'Pushing ${SecondHostRoute.name} then popping-top should present [$FirstPage]',
+  testWidgets('Pushing ${SecondHostRoute.name} then popping-top should present [$FirstPage]',
       (WidgetTester tester) async {
     await pumpRouterApp(tester, router);
     router.push(const SecondHostRoute());
     await tester.pumpAndSettle();
+
     router.popTop();
     await tester.pumpAndSettle();
     expectTopPage(router, FirstRoute.name);
@@ -53,6 +51,7 @@ void main() {
       SecondNested2Route(),
     ]));
     await tester.pumpAndSettle();
+
     expectTopPage(router, SecondNested2Route.name);
     expect(router.urlState.url, '/second/nested2');
   });
@@ -106,8 +105,7 @@ void main() {
     expect(router.urlState.url, '/second');
   });
 
-  testWidgets(
-      'Pushing $SecondHostRoute should add a child router then popping it should remove it',
+  testWidgets('Pushing $SecondHostRoute should add a child router then popping it should remove it',
       (WidgetTester tester) async {
     await pumpRouterApp(tester, router);
     router.push(const SecondHostRoute());